requirements.txt
streamlit
google-genai
tavily-python